Proof of concept about early connection closed detection.
Originally coded against warp 3.3.17 - ported to HEAD for easy diffing.
No changes, but didn't build.

Tests show an understandable race regression:

   Failures:

      test/RunSpec.hs:182:9:
        1) Run, connection termination, IncompleteHeaders
               expected: "Just Warp: Request headers did not finish transmission"
               but got: "Just Warp: Client closed connection prematurely"

                         To rerun use: --match "/Run/connection
                         termination/IncompleteHeaders/"

Generally it would be nicer to not need the recvMsg (which we only use
because it can do MSG_PEEK, which will notify us about a 0-length read
indicating termination, but will not remove data from the buffer in the
normal read case). Rather, maybe warp could use a thread to preemptively
read the socket (and pre-populate the byte source?), so it could notice
the client closing the connection during the course of normal reads. But
I don't know enough about how warp flow of events is constructed, so
this is more a shot in the dark.

Also don't know if the current way introduces performance overhead,
either due to the extra thread, or the possible duplicate reads issued
against the socket (and maybe some need of associated locking in the
kernel).